update BattleService to enable 2 Player Battles via an endpoint & fix bugs in dependency injection

+++ src/main/java/mtcg/model/interfaces/BattleCard.java
@@ -13,5 +13,5 @@ public interface BattleCard extends Card, Attackable {
 
     Effectiveness getEffectiveMultiplier(ElementType otherElementType);
 
-    float getDamageWithEffectiveMultiplier(BattleCard card, ElementType enemyElementType);
+    int getDamageWithEffectiveMultiplier(BattleCard card, ElementType enemyElementType);
 }
